+++ server/controllers/about-the-device-wearer/deviceWearerController.test.ts
@@ -1,11 +1,11 @@
-import { getMockOrder } from '../../test/mocks/mockOrder'
-import { createMockRequest, createMockResponse } from '../../test/mocks/mockExpress'
-import HmppsAuditClient from '../data/hmppsAuditClient'
-import RestClient from '../data/restClient'
-import AuditService from '../services/auditService'
-import DeviceWearerService from '../services/deviceWearerService'
+import { getMockOrder } from '../../../test/mocks/mockOrder'
+import { createMockRequest, createMockResponse } from '../../../test/mocks/mockExpress'
+import HmppsAuditClient from '../../data/hmppsAuditClient'
+import RestClient from '../../data/restClient'
+import AuditService from '../../services/auditService'
+import DeviceWearerService from '../../services/deviceWearerService'
 import DeviceWearerController from './deviceWearerController'
-import TaskListService from '../services/taskListService'
+import TaskListService from '../../services/taskListService'
 
 jest.mock('../services/auditService')
 jest.mock('../services/orderService')
