+++ server/controllers/about-the-device-wearer/deviceWearerController.ts
@@ -1,12 +1,12 @@
 import { Request, RequestHandler, Response } from 'express'
 import z from 'zod'
-import paths from '../constants/paths'
-import { DeviceWearer } from '../models/DeviceWearer'
-import { isValidationResult, ValidationResult } from '../models/Validation'
-import { MultipleChoiceField, TextField } from '../models/view-models/utils'
-import { AuditService, DeviceWearerService } from '../services'
-import { deserialiseDate, getError } from '../utils/utils'
-import TaskListService from '../services/taskListService'
+import paths from '../../constants/paths'
+import { DeviceWearer } from '../../models/DeviceWearer'
+import { isValidationResult, ValidationResult } from '../../models/Validation'
+import { MultipleChoiceField, TextField } from '../../models/view-models/utils'
+import { AuditService, DeviceWearerService } from '../../services'
+import { deserialiseDate, getError } from '../../utils/utils'
+import TaskListService from '../../services/taskListService'
 
 // Basic validation of user submitted form data
 const DeviceWearerFormDataModel = z.object({
