+++ server/controllers/about-the-device-wearer/deviceWearerResponsibleAdultController.ts
@@ -1,13 +1,13 @@
 import z from 'zod'
 import { Request, RequestHandler, Response } from 'express'
-import { DeviceWearerResponsibleAdult } from '../models/DeviceWearerResponsibleAdult'
-import { AuditService } from '../services'
-import { TextField } from '../models/view-models/utils'
-import paths from '../constants/paths'
-import { isValidationResult, ValidationResult } from '../models/Validation'
-import { getError } from '../utils/utils'
-import DeviceWearerResponsibleAdultService from '../services/deviceWearerResponsibleAdultService'
-import TaskListService from '../services/taskListService'
+import { DeviceWearerResponsibleAdult } from '../../models/DeviceWearerResponsibleAdult'
+import { AuditService } from '../../services'
+import { TextField } from '../../models/view-models/utils'
+import paths from '../../constants/paths'
+import { isValidationResult, ValidationResult } from '../../models/Validation'
+import { getError } from '../../utils/utils'
+import DeviceWearerResponsibleAdultService from '../../services/deviceWearerResponsibleAdultService'
+import TaskListService from '../../services/taskListService'
 
 const DeviceWearerResponsibleAdultFormDataModel = z.object({
   action: z.string(),
