+++ server/routes/index.ts
@@ -1,13 +1,13 @@
 import { type RequestHandler, Router } from 'express'
 
-import AttachmentsController from '../controllers/attachmentController'
+import AttachmentsController from '../controllers/attachments/attachmentController'
 import AddressController from '../controllers/contact-information/addressController'
 import ContactDetailsController from '../controllers/contact-information/contactDetailsController'
 import NoFixedAbodeController from '../controllers/contact-information/noFixedAbodeController'
 import InterestedPartiesController from '../controllers/contact-information/interestedPartiesController'
-import DeviceWearerController from '../controllers/deviceWearerController'
-import ResponsibleAdultController from '../controllers/deviceWearerResponsibleAdultController'
-import DeviceWearerCheckAnswersController from '../controllers/deviceWearerCheckAnswersController'
+import DeviceWearerController from '../controllers/about-the-device-wearer/deviceWearerController'
+import ResponsibleAdultController from '../controllers/about-the-device-wearer/deviceWearerResponsibleAdultController'
+import DeviceWearerCheckAnswersController from '../controllers/about-the-device-wearer/deviceWearerCheckAnswersController'
 import InstallationAndRiskController from '../controllers/installationAndRisk/installationAndRiskController'
 import AlcoholMonitoringController from '../controllers/monitoringConditions/alcoholMonitoringController'
 import AttendanceMonitoringController from '../controllers/monitoringConditions/attendanceMonitoringController'
